declarative: add owner ref support
This commit adds an option to setup owner refs for each deployed object.
This ensures that deployed objects are cleaned up as part of the
deletion of the CR.

This is implemented by providing a OwnerSelector that allows the
implementor to specify the owner for each individual object. The most basic
implementation of this is SourceAsOwner which parents all deployment
objects to the root CR.

Flexibility is needed here if the user wants to provide intermediate
owner objects, for example, Addons want this structure:
   Addon CR -> Application -> {Deployment Objects}