Implement Application management
The Application is a CR that represents a given deployment. If a
deployment contains an Application we can specialize it based on the
declarative configuration. We can go further with Addons which have
health/version to expose on the Application.

More info on the application: https://github.com/kubernetes-sigs/application

fixes #5